Move parameter methods to new file.
